Address review feedback for diagnostic infrastructure
- Use PlatformUtil::getEnvironmentVariable instead of raw getenv/_dupenv_s
- Use case-insensitive comparison for diagnostic category matching
- Fix SCOPED_TIMER macro: two-level indirection for __LINE__ expansion
- Fix missing (long long) cast in Windows fprintf for RPC write duration
- Remove deprecated slang.dll copy from VK-GL-CTS nightly workflow
- Add comments explaining drainTestServerStderr safety and stderr logging

+++ source/core/slang-http.cpp
@@ -480,7 +480,7 @@ SlangResult HTTPPacketConnection::write(const void* content, size_t sizeInBytes)
             stderr,
             "[RPC-WRITE] HTTPPacketConnection::write size=%zu duration=%lldus\n",
             sizeInBytes,
-            duration);
+            (long long)duration);
 #endif
 
         // Warn if write took more than 100ms
